+++ docs-src/components/color-palette/ColorPalette.svelte
@@ -6,36 +6,42 @@
 <div class="group">
 	<PaletteBox bg="color-accent" />
 	<PaletteBox bg="color-accent-semi" />
+	<PaletteBox bg="color-accent-gradient" />
 </div>
 
 <h4>Secondary</h4>
 <div class="group">
 	<PaletteBox bg="color-secondary" />
 	<PaletteBox bg="color-secondary-semi" />
+	<PaletteBox bg="color-secondary-gradient" />
 </div>
 
 <h4>Info</h4>
 <div class="group">
 	<PaletteBox bg="color-info" />
 	<PaletteBox bg="color-info-semi" />
+	<PaletteBox bg="color-info-gradient" />
 </div>
 
 <h4>Success</h4>
 <div class="group">
 	<PaletteBox bg="color-success" />
 	<PaletteBox bg="color-success-semi" />
+	<PaletteBox bg="color-success-gradient" />
 </div>
 
 <h4>Warning</h4>
 <div class="group">
 	<PaletteBox bg="color-warning" />
 	<PaletteBox bg="color-warning-semi" />
+	<PaletteBox bg="color-warning-gradient" />
 </div>
 
 <h4>Danger</h4>
 <div class="group">
 	<PaletteBox bg="color-danger" />
 	<PaletteBox bg="color-danger-semi" />
+	<PaletteBox bg="color-danger-gradient" />
 </div>
 
 <h3>Highlight colors</h3>
@@ -43,6 +49,7 @@
 	<PaletteBox bg="color-highlight" />
 	<PaletteBox bg="color-highlight-semi" />
 	<PaletteBox bg="color-highlight-1" />
+	<PaletteBox bg="color-highlight-gradient" />
 </div>
 
 <h3>Background colors</h3>
@@ -61,6 +68,7 @@
 	<PaletteBox text="color-text-semi" />
 	<PaletteBox text="color-text-1" />
 	<PaletteBox text="color-text-2" />
+	<PaletteBox text="color-background" bg="color-text-gradient" />
 </div>