+++ docs-src/components/color-palette/ColorPaletteBox.svelte
@@ -1,15 +1,16 @@
-<div class="palette-box" style="background-color: var(--ui-{bg});color: var(--ui-{text});">
+<div class="palette-box" style="background: var(--ui-{bg}); color: var(--ui-{text});">
 	background: --ui-{bg}<br>
 	text: --ui-{text}
 </div>
 
-<script>
-/**
- * @typedef {Object} Props
- * @property {string} [bg]
- * @property {string} [text]
- */
+<script lang="ts">
+interface Props {
+	bg?: string;
+	text?: string;
+}
 
-/** @type {Props} */
-const { bg = 'color-background', text = 'color-text' } = $props();
+const {
+	bg = 'color-background',
+	text = 'color-text'
+}: Props = $props();
 </script>