+++ CHANGELOG.md
@@ -33,6 +33,7 @@ Thousand thanks to the Svelte's team, for the [long months of hard work](https:/
 
 
 ### Other changes
+- Added some gradients to the palette.
 - Added `ArrowLeft` and `ArrowRight` to `Toggle` component, to allow toggling the value with the keyboard.
 - Added `ArrowUp` and `ArrowDown` to `InputNumber` component, to allow incrementing/decrementing the value with the keyboard.
 - Fixed `Tooltip`: when mouse entered target from the top - tooltip would show up and hide instantly.